[less] Adapted batch compiler so it can be used for rollups.
The batch compiler will now recursively parse all .less files found
under the source directory.  You can also pass in the '-C' argument
specifying the exact list of files which are to be compiled to .css.

So the rollups compiler can defer all less compilation to the batch
compiler, passing in the unique list of all .less -> .css file
transforms.  Then it only has to concatenate the resulting .css files
to produce the final rolled-up file.