[less] Backed out adv. pct. operations in favor of compat.
Rationale: I initially implemented changes to how percentages behave in
operations, but this breaks compatibility.  I have very little feedback
on this change, and the rollout could be complex, given the number of
diffs it creates against the ~1200 developer sites, and the complexity
involved with rolling out a change that breaks compatibility with
less.js.  Given these concerns I decided to back out this change, but
leave the code intact for future consideration.