[less] Compress mode output for repr().
In production, we have cases where we need to minify CSS but do not want
to evaluate any of the rules.  This change allows us to parse a
stylesheet, which may be CSS or LESS, and emit a compressed
representation of it without doing any evaluation.

Originally NodeRenderer was intended only for CSS output, with
Node.repr() for canonical LESS output.  They should be merged since
there is significant overlap, now that both subsystems support compress
(minify) output.