Operation strict math / parens logic.
This removes the need for a special parselet for the 'font' rule,
whose shorthand resembles a division operation.

When the 'font' rule is detected, the parser switches to strict math
mode, which informs all nested operations that they're inside a strict
math expression.

The parser also communicates whether the operation is inside nested
parenthesis.  During eval() and repr(), this information is used by the
operation to return the correct value.