Render selector mixin paths post-evaluation.
When a ruleset is defined, one or more of its selectors may allow it
to be used as a mixin.

Before the ruleset is evaluated, only a selector that requires no
evaluation, and consists of only text elements can be used as a mixin
path.  Example:  ".foo".

Since selectors can contain embedded variables, a selector such as
".@{name}" may be used as a selector post-evaluation.  If @name ==
"bar", the rendered mixin path would be [".bar"] for this ruleset.