Imports with path variable interpolation.
First pass at wiring up post-parse evaluation of imports whose paths use
variables.  This requires effectively doing an evaluation of all of
these import statements.

To optimize, we capture the relevant imports during the parse by
recording the entry / exit of each nested block, and saving this as
a closure whenever a block containing an interpolated import is
detected.