Support dynamic imports, new hierarchical parser.
In the 1.x branch, parsing of imports was done in a separate context,
cut off from the parent parse.  This means that nested imports with
variable interpolation would not (easily) be able to resolve against the
parent parse context.

In 2.x this is changed so that multiple parses can be run against the
same parse context.  When an IMPORT statement is encountered, the
current stream will be pushed onto the stack and a new stream will be
created.  Once the imported stream reaches EOF, the stack will be popped
and the top-level parse will resume.

Additionally, when an import is encountered whose path requires variable
interpolation, we defer the block and a closure onto a "deferred
evaluation list".  Once the parse completes any entries in this list
will be scanned and evaluated.

This results in a much faster parse, since it avoids having to scan the
entire tree recursively looking for IMPORT statements and expanding
them post-parse.  For deeply-nested stylesheets containing many imports
this can require many scans of the tree to find and expand every
static and dynamic import.

Also added test cases for dynamic import and comments.

Example of imports with interpolated paths:

 global.less:

     @import "child-@{name1}.less";
     @import "vars.less";

 vars.less:

     @name1: 'foo';
     @name2: 'bar';

 child-foo.less:

     .ruleset-foo {
        color: blue;
    }
     @import 'child-@{name2}.less';

 child-bar.less:

     .ruleset-bar {
        font-size: 10px;
    }