Selector canonicalization, pattern recognizers.
1. Selector canonicalization

Selectors can contain embedded variables.  Depending on the way variables are
embedded, it can produce what are logically equivalent selectors with very
different syntax trees. This makes comparison very difficult without some form
of normalization.

Now a selector that requires evaluation will be rendered and parsed to ensure
its syntax tree is consistent, making it possible to accurately compare
selectors regardless of the complexity of their parse/evaluation trees.

Canonicalization is a requirement of the extend feature.  Selectors that have
embedded variables are difficult to compare otherwise, as the variable may hold
a rich set of partial combinators and elements inside escaped strings:

    @name: ~'foo.bar[a="b"]';
    .@{name} { ... }          // A
    .foo.bar[a="b"] { ... }   // B

Selectors A and B are logically equivalent but need to be canonicalized in
order to be compared using equals(), due to the way their syntax trees differ
post-evaluation.

2. Pattern recognizer state machine and unit tests.

In the first phase these recognizers will be used to render and parse
selectors, to produce a canonical form for comparison and search / replace.

This pattern recognizer subsytem to eventually replace the use of
java.util.regex-based pattern matchers.

The recognizers run faster than their equivalent Java regex, in some cases by a
factor of 2 to 5, which will speed up overall parsing and reduce the number of
fields and methods the LessStream instances hold.