Combinators split out from elements.
Introduced new generic SelectorPart as base class for all nodes which
can be part of a Selector.

Previously combinators were attached to the Element nodes. This adds
complexity when performing searches and combining selectors together.

Now combinators are standalone nodes, so they can more easily be
prepended / appended to selectors, and optionally filtered out.  This
also removes the need to have a NULL combinator for elements that are
attached to one another.