Generate a JSON representation of a parsed stylesheet.
This builds a JSON representation of a stylesheet. The intent is to use
this to bypass LESS parsing in the browser, instead using the faster
native JSON parse and directly translating this to a LESS instruction
tree.

All string references are interned and replaced with a unique
identifier, compressing the often-redundant string information.  For
example the string 'px' may occur many times, e.g. '10px', '12px'. So
we store the string 'px' once and reference it with a unique id. When
the JSON is expanded and the instruction tree constructed, all
references to 'px' will point to the same string instance, saving
memory.

To save additional memory, we also add an optimization pass over the
string table which sorts it by frequency, so more frequently-occurring
strings get lower id numbers, which can be represented by fewer bytes.
So if 'px' occurs the most it would be represented by the id '0'.

Currently we do not encode line/char offset information directly in the
instruction tree, since this is only used in rare instances when an
error occurs. Instead, as instructions are emitted, we use their
implicit index to emit a Base64 VLQ-encoded line/char offset, as well as
an index to the path information for block nodes. This structure can
then be used to access position information at runtime if an error
occurs.

This position information is currently optional as it bloats up the
size of the JSON output by approx 50%.