Optimization of mixin resolution.
Maintain an index attached to each block that maps a mixin path prefix
to the list of mixins that share that prefix. This reduces the search
space considerably when doing mixin resolution, as the vast majority
of blocks will not contain a match, but would otherwise require
scanning all nodes inside the block and performing path prefix
comparisons on each mixin definition found.