Replace LinkedHashSet with custom set in CssModel
Benchmarking has shown significant overhead in maintaining the ordered
set of nodes within a CssBlock. Our use case requires that nodes be
unique, insertion order maintained, and if a duplicate node is added
it moves to the end of the block.

Using LinkedHashSet for this requires the following code:

    // if exists, remove duplicate. otherwise dupe would
    // stay at original insertion point
    if (nodes.contains(n)) {
        nodes.remove(n);
    }
    // add to end of ordered set
    nodes.add(n);

We replace this code with a more efficient add/update/move-to-end in
a single operation.